pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "label-generator"
version = "0.1.0"
description = "Batch apparel label PNG generator with JAN-13 barcode support"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0.0",
    "python-barcode>=0.15.1",
    "pandas>=2.0.0",
    "typer>=0.12.0",
    "openpyxl>=3.1.0",
]

[project.scripts]
label-generator = "label_generator.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
